requirements.txt
numpy
torch
pandas
pyyaml
tensorboard
matplotlib